move sync loop tests to sync_loop_tests.rs
